Preliminary stage/requirements.txt
django==1.9
PyGithub
beautifulsoup4
